fix(architect-web): keep top nav mounted across /protocol/* tabs
