+++ apps/architect-web/src/components/ProjectNav/ProjectNav.tsx
@@ -1,32 +1,17 @@
 import {
   BookOpenText,
-  Check,
-  Download,
   FileImage,
-  Home,
   type LucideIcon,
   Printer,
-  Redo,
   Timeline,
-  Undo,
 } from 'lucide-react';
 import { motion } from 'motion/react';
-import type React from 'react';
-import { useCallback, useEffect, useState } from 'react';
 import { useSelector } from 'react-redux';
 import { Link, useLocation } from 'wouter';
 
-import Tooltip from '~/components/NewComponents/Tooltip';
-import { useAppDispatch } from '~/ducks/hooks';
-import { redo, undo } from '~/ducks/modules/activeProtocol';
-import { actionCreators as dialogActions } from '~/ducks/modules/dialogs';
-import { exportNetcanvas } from '~/ducks/modules/userActions/userActions';
-import { useReturnToStartDialog } from '~/hooks/useReturnToStartDialog';
-import Button, { IconButton } from '~/lib/legacy-ui/components/Button';
-import { getCanRedo, getCanUndo, getProtocolName } from '~/selectors/protocol';
+import { getProtocolName } from '~/selectors/protocol';
 import { cx } from '~/utils/cva';
 
-import ActionToolbar from './ActionToolbar';
 import Breadcrumb, { type BreadcrumbItem } from './Breadcrumb';
 import NavShell from './NavShell';
 
@@ -43,50 +28,9 @@ const TABS: Tab[] = [
   { href: '/protocol/summary', label: 'Summary', Icon: Printer },
 ];
 
-type ProjectNavProps = {
-  extraActions?: React.ReactNode;
-};
-
-const ProjectNav = ({ extraActions }: ProjectNavProps) => {
+const ProjectNav = () => {
   const [location] = useLocation();
-  const dispatch = useAppDispatch();
   const protocolName = useSelector(getProtocolName);
-  const canUndo = useSelector(getCanUndo);
-  const canRedo = useSelector(getCanRedo);
-  const handleReturnToStart = useReturnToStartDialog();
-
-  const [isExporting, setIsExporting] = useState(false);
-  const [downloadSuccess, setDownloadSuccess] = useState(false);
-
-  const handleUndo = useCallback(() => dispatch(undo()), [dispatch]);
-  const handleRedo = useCallback(() => dispatch(redo()), [dispatch]);
-
-  const handleDownload = useCallback(async () => {
-    try {
-      setIsExporting(true);
-      await dispatch(exportNetcanvas()).unwrap();
-      setDownloadSuccess(true);
-    } catch (error) {
-      dispatch(
-        dialogActions.openDialog({
-          type: 'Error',
-          title: 'Failed to export protocol',
-          message: error instanceof Error ? error.message : String(error),
-        }),
-      );
-    } finally {
-      setIsExporting(false);
-    }
-  }, [dispatch]);
-
-  useEffect(() => {
-    if (!downloadSuccess) return;
-    const timer = setTimeout(() => setDownloadSuccess(false), 2000);
-    return () => clearTimeout(timer);
-  }, [downloadSuccess]);
-
-  // Summary is read-only: undo/redo would have no visible effect.
-  const isSummary = location === '/protocol/summary';
 
   const breadcrumbItems: BreadcrumbItem[] = [
     { label: protocolName ?? 'Untitled protocol' },
@@ -128,51 +72,10 @@ const ProjectNav = ({ extraActions }: ProjectNavProps) => {
   );
 
   return (
-    <>
-      <NavShell
-        leading={<Breadcrumb items={breadcrumbItems} />}
-        trailing={tabs}
-      />
-      <ActionToolbar>
-        <Button onClick={handleReturnToStart} color="platinum" icon={<Home />}>
-          Return to start
-        </Button>
-        {extraActions}
-        {!isSummary && (
-          <>
-            <IconButton
-              variant="text"
-              icon={<Undo />}
-              onClick={handleUndo}
-              disabled={!canUndo}
-              aria-label="Undo"
-            />
-            <IconButton
-              variant="text"
-              icon={<Redo />}
-              onClick={handleRedo}
-              disabled={!canRedo}
-              aria-label="Redo"
-            />
-          </>
-        )}
-        <Tooltip content="Download .netcanvas protocol">
-          <Button
-            onClick={handleDownload}
-            color="sea-green"
-            content={
-              downloadSuccess
-                ? 'Downloaded'
-                : isExporting
-                  ? 'Downloading...'
-                  : 'Download'
-            }
-            disabled={isExporting}
-            icon={downloadSuccess ? <Check /> : <Download />}
-          />
-        </Tooltip>
-      </ActionToolbar>
-    </>
+    <NavShell
+      leading={<Breadcrumb items={breadcrumbItems} />}
+      trailing={tabs}
+    />
   );
 };
