+++ apps/architect-web/src/components/Protocol.tsx
@@ -1,5 +1,6 @@
 import useProtocolLoader from '~/hooks/useProtocolLoader';
 
+import ProjectActions from './ProjectNav/ProjectActions';
 import ProtocolInfoCard from './ProtocolInfoCard';
 import Timeline from './Timeline';
 
@@ -7,6 +8,7 @@ const Protocol = () => {
   useProtocolLoader();
   return (
     <div className="mt-(--space-xl) flex flex-col items-center">
+      <ProjectActions showReturnToStart />
       <ProtocolInfoCard />
       <Timeline />
     </div>
