+++ apps/architect-web/src/components/pages/AssetsPage.tsx
@@ -2,6 +2,7 @@ import AssetBrowser from '~/components/AssetBrowser';
 import { Layout } from '~/components/EditorLayout';
 import ExternalLink from '~/components/ExternalLink';
 import PageHeading from '~/components/ProjectNav/PageHeading';
+import ProjectActions from '~/components/ProjectNav/ProjectActions';
 import useProtocolLoader from '~/hooks/useProtocolLoader';
 
 const AssetsPage = () => {
@@ -10,6 +11,7 @@ const AssetsPage = () => {
 
   return (
     <Layout>
+      <ProjectActions />
       <PageHeading
         title="Resource Library"
         description={
